+++ branches/TaksoPark1.1/DBFiles/CreateLocalDB.sql
@@ -13,7 +13,7 @@ create table [Request]
 	[FinishPoint] nvarchar(512),
 	[OperatorId] int,
 	[DriverId] int,
-	[Price] int not null,
+	[Price] numeric(18,4) not null,
 	[Additional] nvarchar(2048)
 )
 
@@ -651,7 +651,7 @@ CREATE PROCEDURE sp_CreateRequest
    @FinishPoint nvarchar(512),
    @OperatorId int,
    @DriverId int,
-   @Price int,
+   @Price numeric(18,4),
    @Additional nvarchar(2048),
    @RequestId int = NULL OUTPUT
 AS
@@ -787,7 +787,7 @@ CREATE PROCEDURE sp_UpdateRequest
     @FinishPoint nvarchar(512),
     @RequestId int,
     @DriverId int,
-    @Price int,
+    @Price numeric(18,4),
     @OperatorId int,
     @Additional nvarchar(2048)
 AS
